Cleaning up
cleaned everything up

+++ Titouan tests/generate_graph.h
@@ -1,6 +1,6 @@
 #ifndef GENERATE_GRAPH_H
 #define GENERATE_GRAPH_H
 
-int generate_graph(size_t height, size_t width);
+int generate_graph(int height, int width);
 
 #endif
